Submission Version
Added comments (sparse).

Added initial tasks that provide some  clues as to how the interface
works.

Made it so that when a certain category is being displayed, the order
updates when the sort is changed (before, it stayed in the order the
sort that was being displayed when the user decided to only show the
category)